ParanaTreeAI_TaskDay-01/requirement.txt
pandas
numpy
openai
python-dotenv
ipykernel